address review: share embedded-entry resolution with Worker, scrub fork env var from process.env
The extension-remapping probe now lives as a provided method on the
resolver's StandaloneModuleGraph trait so new Worker(path) and
child_process.fork(path) inside a compiled binary use the same table.

BUN_INTERNAL_FORK_ENTRY is removed from the env loader map alongside
NODE_CHANNEL_FD so it never shows up in process.env. std::env::remove_var
alone does not touch the WTF-8 snapshot on Windows.

The compile/ChildProcessForkEmbedded test now reads stdout and the IPC
message from state inside the 'close' handler so the assertion does not
depend on message vs exit ordering.
